Add contract and device registration features
- Updated `ContractService.cs` to include `ContractId`.
- Modified `PaymentContractResponse.cs` with new properties.
- Created `RegisterDeviceRequest.cs` for device registration.
- Cleaned up `SCPM_Data.sql` by removing sensitive entries.
- Added new `FCM` table in `SCPM_Database.sql`.
- Introduced `FcmController.cs` for handling device registration.
- Updated solution file `ScpmaBe.WebApi.sln` for new structure.
- Added cleanup script in `SCPM_Database_Cleanup.sql`.

+++ ScpmaBe.Services/ParkingSpaceService.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using ScpmaBe.Repositories;
 using ScpmaBe.Repositories.Entities;
 using ScpmaBe.Repositories.Interfaces;
+using ScpmaBe.Services.Enums;
 using ScpmaBe.Services.Interfaces;
 using ScpmaBe.Services.Models;
 using ScpmBe.Services.Exceptions;
@@ -10,10 +12,15 @@ namespace ScpmaBe.Services
     public class ParkingSpaceService : IParkingSpaceService
     {
         private readonly IParkingSpaceRepository _parkingSpaceRepository;
+        private readonly IParkingStatusSensorRepository _parkingStatusSensorRepository;
+        private readonly IFloorRepository _floorRepository;
 
-        public ParkingSpaceService(IParkingSpaceRepository parkingspaceRepository)
+        public ParkingSpaceService(
+            IParkingSpaceRepository parkingspaceRepository, IParkingStatusSensorRepository parkingStatusSensorRepository, IFloorRepository floorRepository)
         {
             _parkingSpaceRepository = parkingspaceRepository;
+            _parkingStatusSensorRepository = parkingStatusSensorRepository;
+            _floorRepository = floorRepository;
         }
 
         public async Task<List<ParkingSpace>> GetPaging(int pageIndex, int pageSize)
@@ -21,39 +28,74 @@ namespace ScpmaBe.Services
             return await _parkingSpaceRepository.GetAll().Skip((pageIndex - 1) * pageSize).Take(pageSize).ToListAsync();
         }
 
-        public async Task<ParkingSpace> GetById(int id)
+        public async Task<ParkingSpaceResponse> GetById(int id)
         {
             var parkingspace = await _parkingSpaceRepository.GetById(id);
 
             if (parkingspace == null) throw AppExceptions.NotFoundParkingSpace();
 
-            return parkingspace;
+            return new ParkingSpaceResponse
+            {
+                ParkingSpaceId = parkingspace.ParkingSpaceId,
+                FloorId = parkingspace.FloorId,
+                ParkingSpaceName = parkingspace.ParkingSpaceName,
+                Status = ((ParkingSpaceStatus)parkingspace.Status).ToString()
+            };
         }
 
-        public async Task<ParkingSpace> AddParkingSpaceAsync(AddParkingSpaceRequest request)
+        public async Task<ParkingSpaceResponse> AddParkingSpaceAsync(AddParkingSpaceRequest request)
         {
             var newParkingSpace = new ParkingSpace
             {
                 ParkingSpaceName = request.ParkingSpaceName,
                 FloorId = request.FloorId,
-                Status = request.Status,
+                Status = 1,
             };
 
-            return await _parkingSpaceRepository.Insert(newParkingSpace);
+            var floor = await _floorRepository.GetById(request.FloorId);
+
+            if (floor == null) throw AppExceptions.NotFoundFloor();
+
+            var nameExisted = await _parkingSpaceRepository.GetAll().AnyAsync(x => x.ParkingSpaceName.ToLower() == request.ParkingSpaceName.ToLower() && x.FloorId == request.FloorId);
+
+            if(nameExisted) throw AppExceptions.ParkingSpaceNameExisted();
+
+            floor.TotalParkingSpace++;
+
+            var addedEntity = await _parkingSpaceRepository.Insert(newParkingSpace);
+
+            return new ParkingSpaceResponse
+            {
+                ParkingSpaceId = addedEntity.ParkingSpaceId,
+                FloorId = addedEntity.FloorId,
+                ParkingSpaceName = addedEntity.ParkingSpaceName,
+                Status = ((ParkingSpaceStatus)addedEntity.Status).ToString()
+            };
         }
 
-        public async Task<ParkingSpace> UpdateParkingSpaceAsync(UpdateParkingSpaceRequest request)
+        public async Task<ParkingSpaceResponse> UpdateParkingSpaceAsync(UpdateParkingSpaceRequest request)
         {
             var updateParkingSpace = await _parkingSpaceRepository.GetById(request.ParkingSpaceId);
-           
+
             if (updateParkingSpace == null) throw AppExceptions.NotFoundParkingSpace();
 
+            var floorId = updateParkingSpace.FloorId;
+
+            var nameExisted = await _parkingSpaceRepository.GetAll().AnyAsync(x => x.ParkingSpaceName.ToLower() == request.ParkingSpaceName.ToLower() && x.FloorId == floorId && x.ParkingSpaceId != request.ParkingSpaceId);
+
+            if(nameExisted) throw AppExceptions.ParkingSpaceNameExisted();
+
             updateParkingSpace.ParkingSpaceName = request.ParkingSpaceName;
-            updateParkingSpace.Status = request.Status;
 
             await _parkingSpaceRepository.Update(updateParkingSpace);
 
-            return updateParkingSpace ;
+            return new ParkingSpaceResponse
+            {
+                ParkingSpaceId = updateParkingSpace.ParkingSpaceId,
+                FloorId = updateParkingSpace.FloorId,
+                ParkingSpaceName = updateParkingSpace.ParkingSpaceName,
+                Status = ((ParkingSpaceStatus)updateParkingSpace.Status).ToString()
+            };
         }
 
         public async Task<bool> DeleteParkingSpaceAsync(int id)
@@ -63,6 +105,12 @@ namespace ScpmaBe.Services
                 var parkinglot = await _parkingSpaceRepository.GetById(id);
 
                 if (parkinglot == null) throw AppExceptions.NotFoundParkingSpace();
+
+                var floor = await _floorRepository.GetById(parkinglot.FloorId);
+
+                if (floor == null) throw AppExceptions.NotFoundFloor();
+
+                floor.TotalParkingSpace--;
 
                 await _parkingSpaceRepository.Delete(id);
 
@@ -74,9 +122,49 @@ namespace ScpmaBe.Services
             }
         }
 
-        public async Task<List<ParkingSpace>> GetParkingSpacesByFloor(int floorId)
+        public async Task<List<ParkingSpaceResponse>> GetParkingSpacesByFloor(int floorId)
         {
-            return await _parkingSpaceRepository.GetAll().Where(x => x.FloorId == floorId).ToListAsync();
+            return await _parkingSpaceRepository
+                            .GetAll()
+                            .Where(x => x.FloorId == floorId)
+                            .Select(x => new ParkingSpaceResponse
+                            {
+                                ParkingSpaceId = x.ParkingSpaceId,
+                                ParkingSpaceName = x.ParkingSpaceName,
+                                FloorId = x.FloorId,
+                                Status = ((ParkingSpaceStatus)x.Status).ToString()
+                            }).ToListAsync();
+        }
+
+        public async Task<bool> ChangeStatus(string apiKey)
+        {
+            var parkingStatusSensor = await _parkingStatusSensorRepository.GetAll().FirstOrDefaultAsync(x => x.ApiKey == apiKey);
+
+            if (parkingStatusSensor == null) return false;
+
+            try
+            {
+                var parkingSpace = await _parkingSpaceRepository.GetById(parkingStatusSensor.ParkingSpaceId);
+
+                if (parkingSpace == null) return false;
+
+                if(parkingSpace.Status == (int)ParkingSpaceStatus.Occupied)
+                {
+                    parkingSpace.Status = (int)ParkingSpaceStatus.Pending;
+                }
+                else if(parkingSpace.Status == (int)ParkingSpaceStatus.Pending)
+                {
+                    parkingSpace.Status = (int)ParkingSpaceStatus.Occupied;
+                }
+
+                await _parkingSpaceRepository.Update(parkingSpace);
+            }
+            catch
+            {
+                return false;
+            }
+
+            return true;
         }
     }
 }
